Third and last part of transforming the searcher to the final state - added input mechanics and improved the design
