+++ package.json
@@ -17,6 +17,7 @@
     "@types/react-dom": "^17.0.11",
     "@types/react-loadable": "^5.5.6",
     "@types/react-loadable-visibility": "^3.0.2",
+    "@types/react-pdf": "^5.7.2",
     "@types/react-redux": "^7.1.22",
     "@types/react-router-dom": "^5.3.3",
     "@types/react-window": "^1.8.5",
