+++ src/components/helperComponents/searcher/searchResultComponent.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { SearchingMainResult, SearchingResultHeader, SearchingResultSubInfo, 
-    SearchingResultTagsSection, SearchingResultTag } from "../../../styled/subpages/searcher";
+    SearchingResultTagsSection, SearchingResultTag } from "../../../styled/subpages/searcher/searcherResults";
 
 interface SearchResultComponentInterface{
     align: boolean,
